Move starter_cards tests to tests/ directory
Move inline unit tests from src/starter_cards.rs to
tests/starter_cards_test.rs as integration tests, following
the project convention of keeping tests in the tests/ folder.
